requirements/dev.txt
-r observability.txt

amqp==5.3.1
asgiref==3.11.1
bcrypt==4.1.3
billiard==4.2.4
celery==5.3.6
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.0.4
gunicorn==22.0.0
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.15.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0


drf-spectacular==0.27.2

# CI / QA tooling
bandit==1.7.10
flake8==7.1.1
mypy==1.13.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
django-migration-linter==5.2.0
firebase-admin
